+++ src/scripts/process/tasks/scrapping.tab.js
@@ -2,7 +2,7 @@ import {logWelcome} from "../../units/utils/consoleMessages.all";
 import appInfos from "../../../infos.json";
 import {defineExtractor} from "../../dispatchers/defineExtractor";
 import {defineExportMethod} from "../../dispatchers/defineExportMethod";
-import {updateClickIconCount} from "../storage/clickIconCountContext.tab";
+import {updateClickIconCount} from "../../units/internals/icon/clickIconCountContext.bg";
 
 /**
  * @description - Launch the export process
