+++ src/scripts/process/tasks/setDisplayModalOnInstall.bg.js
